Add DTOs for deposit, stock creation, and trade orders; refactor holdings and user entities.
Big refactors to backend

+++ modules/domain/src/main/kotlin/dev/freddiesilver/stocksim/trading/tradeorder/TradeOrder.kt
@@ -9,7 +9,7 @@ data class TradeOrder(
     val user: User,
     val stock: Stock,
     val type: OrderType,
-    val quantity: Int,
+    val quantity: BigDecimal,
     val priceValueAtOrder: BigDecimal,
     var status: OrderStatus,
 )
